requirements.txt
flickrapi~=2.4.0
pandas~=2.2.2
python-dotenv~=1.2.2